removed int rounding of text dimensions added exception if ypos-calc-loop runs more than 1000 times

+++ trunk/Baselet/src/com/baselet/control/Utils.java
@@ -258,7 +258,7 @@ public abstract class Utils {
 			s = s.replace(' ', '.');
 		}
 		TextLayout tl = new TextLayout(s, font, frc);
-		return new DimensionDouble((int) tl.getBounds().getWidth(), (int) tl.getBounds().getHeight());
+		return new DimensionDouble(tl.getBounds().getWidth(), tl.getBounds().getHeight());
 	}
